Refactor BungeeCord support to more generic 'Proxy' support
