+++ src/java6/java/org/spongepowered/mod/SpongeCoremod.java
@@ -81,8 +81,9 @@ public class SpongeCoremod implements IFMLLoadingPlugin {
         // Let's get this party started
         SpongeLaunch.setupMixinEnvironment()
                 .addConfiguration("mixins.forge.core.json")
-                .addConfiguration("mixins.forge.bungeecord.json")
                 .addConfiguration("mixins.forge.entityactivation.json")
+                .addConfiguration("mixins.forge.entityactivation.json")
+                .addConfiguration("mixins.forge.proxy.json")
                 .registerTokenProviderClass("org.spongepowered.mod.SpongeCoremod$TokenProvider");
 
         // Add pre-init mixins
