--- src/main/resources/mixins.forge.bungeecord.json
@@ -1,12 +1,0 @@
-{
-  "minVersion": "0.5.1",
-  "package": "org.spongepowered.mod.mixin.bungee",
-  "plugin": "org.spongepowered.mod.mixin.plugin.bungee.BungeePlugin",
-  "refmap": "mixins.forge.refmap.json",
-  "mixins": [
-    "network.MixinNetHandlerHandshakeTCP"
-  ],
-  "injectors": {
-    "defaultRequire": 1
-  }
-}
