+++ src/main/resources/mixins.forge.proxy.json
@@ -1,0 +1,12 @@
+{
+    "minVersion": "0.5.1",
+    "package": "org.spongepowered.mod.mixin.proxy",
+    "plugin": "org.spongepowered.mod.mixin.plugin.proxy.ProxyPlugin",
+    "refmap": "mixins.forge.refmap.json",
+    "mixins": [
+        "network.MixinNetHandlerHandshakeTCP"
+    ],
+    "injectors": {
+        "defaultRequire": 1
+    }
+}
